feat(cairo): add telemetry for SQL compiler + CairoEngine events

+++ core/src/main/java/io/questdb/griffin/CompiledQueryImpl.java
@@ -32,7 +32,7 @@ public class CompiledQueryImpl implements CompiledQuery {
     private RecordCursorFactory recordCursorFactory;
     private InsertStatement insertStatement;
     private TextLoader textLoader;
-    private int type;
+    private short type;
 
     @Override
     public RecordCursorFactory getRecordCursorFactory() {
@@ -50,7 +50,7 @@ public class CompiledQueryImpl implements CompiledQuery {
     }
 
     @Override
-    public int getType() {
+    public short getType() {
         return type;
     }
